The square and perimetr

+++ OOP/src/Circle.java
@@ -4,12 +4,12 @@ public class Circle implements SquareAndPerimetr {
     double p;
 
     public void Square() {
-        s = (double) 3.14 * r * r;
+        s = 3.14 * r * r;
         System.out.println("The square of the Circle= " + s);
     }
 
     public void Perimetr() {
-        p = (double) 2 * 3.14 * r;
+        p = 2 * 3.14 * r;
         System.out.println("The perimetr of the Circle= " + p);
     }
 }
